Check if instance exists before attempting deletion
Also remove password from response bodies regarding user passwords

+++ packages/backend/src/controllers/client.js
@@ -101,6 +101,11 @@ const deleteById = async (req, res, next) => {
   console.log('*************************');
   try {
     const client = await Client.findByPk(req.params.clientId);
+    if (!client) {
+      return res
+        .status(404)
+        .json({ error: { title: 'Client not found' }, data: null });
+    }
     await client.destroy();
     return res.status(204).send();
   } catch (err) {
